spec: move printing specs behind feature flag
